pyproject.toml
[tool.poetry]
name = "shop_bot"
version = "0.2.0"
description = ""
authors = ["MansorY23 <arsa20034@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.0"
sqlalchemy = "^2.0.34"
redis = "^5.0.8"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
